+++ AlphaEngine/Level/LogoLevel.cpp
@@ -28,28 +28,28 @@ void level::LogoLevel::Init()
 
 void level::LogoLevel::Update() 
 {
-	float dt = AEFrameRateControllerGetFrameTime();
+	f64 dt = AEFrameRateControllerGetFrameTime();
 
 	timer += dt;
 
 	if (timer < 3.f)
 	{
-		digipenLogo->GetComponent<SpriteComp>()->SetAlpha(timer / 3.f);
+		digipenLogo->GetComponent<SpriteComp>()->SetAlpha((float)timer / 3.f);
 	}
 
 	else if (timer < 5.f)
 	{
-		digipenLogo->GetComponent<SpriteComp>()->SetAlpha(1 - (timer - 3.f) / 2.f);
+		digipenLogo->GetComponent<SpriteComp>()->SetAlpha(1 - ((float)timer - 3.f) / 2.f);
 	}
 
 	else if (timer < 8.f)
 	{
-		teamLogo->GetComponent<SpriteComp>()->SetAlpha((timer - 5.f) / 3.f);
+		teamLogo->GetComponent<SpriteComp>()->SetAlpha(((float)timer - 5.f) / 3.f);
 	}
 
 	else if (timer < 10.f)
 	{
-		teamLogo->GetComponent<SpriteComp>()->SetAlpha(1 - (timer - 8.f) / 2.f);
+		teamLogo->GetComponent<SpriteComp>()->SetAlpha(1 - ((float)timer - 8.f) / 2.f);
 	}
 
 	else
